Move blocks API endpoints to their own PHP file
This allows us to clean up the functions.php file a bit
