+++ src/Blocks/Articles.php
@@ -8,6 +8,8 @@
 
 namespace P4\MasterTheme\Blocks;
 
+use WP_REST_Server;
+
 /**
  * Class Articles
  *
@@ -108,6 +110,7 @@ class Articles extends BaseBlock
 
         add_action('enqueue_block_editor_assets', [ self::class, 'enqueue_editor_assets' ]);
         add_action('wp_enqueue_scripts', [ self::class, 'enqueue_frontend_assets' ]);
+        add_action('rest_api_init', [ self::class, 'register_endpoint' ]);
     }
 
     /**
@@ -415,5 +418,30 @@ class Articles extends BaseBlock
         }
 
         return $args;
+    }
+
+    /**
+     * Register endpoint to retrieve the articles for the Articles block.
+     *
+     * @example GET /wp-json/planet4/v1/get-posts/
+     */
+    public static function register_endpoint(): void
+    {
+        register_rest_route(
+            self::REST_NAMESPACE,
+            'get-posts',
+            [
+                [
+                    'permission_callback' => static function () {
+                        return true;
+                    },
+                    'methods' => WP_REST_Server::READABLE,
+                    'callback' => static function ($request) {
+                        $covers = self::get_posts($request->get_params());
+                        return rest_ensure_response($covers);
+                    },
+                ],
+            ]
+        );
     }
 }
